Fix karma.conf.js to only watch test.js
Karma was previously configured to watch all of the files in the pattern
which meant that any time those files changed (the autobuilder copies
them over), it triggered a rerun. This was usually before the test file
was ready to run, which caused test failures or other glitches.

This commit fixes the Karma config so that autorun test only run when
compilation has finished.