Extract event from coeffects in the same place as in db-handler->interceptor
A good principle to follow in designing anything is to make things that are the
same look the same, and things that behave differently look differently.

`db-handler->interceptor` and `fx-handler->interceptor` are very similar. Their
only difference is that the latter calls the `handler-fn` with all coeffects and
assoces the result to the `:effects` key directly, whereas the former calls the
`handler-fn` with just the db from inside the coeffects, and assoces the result
to just the db in the effects.

However the code of these two functions slightly differed in *where* the event
and the db resp. the whole coeffect was extracted from the context. Eliminate
that difference so that it is clearer what's different and what's the same about
these two functions.