Simplify cljdoc config
We have, and will continue to, put a lot of time into maintaining
re-frame's official documentation site at
http://day8.github.io/re-frame/

Anything we do with cljdoc.edn will always not quite reflect the correct
or intended styling, navigation etc of the official documentation site.

It is also duplication of effort for every change we would need to
modify both documentation configs etc.

As such we have decided that we will not maintain cljdoc documentation
and instead disable it and/or redirect it to the official documentation
site as much as technically possible.